refactor(query): move to read mod
