+++ src/query/storages/fuse/src/operations/mod.rs
@@ -32,7 +32,7 @@ mod fuse_parquet_source;
 mod fuse_source;
 mod revert;
 pub mod util;
-mod fuse_parquet_source_new;
+mod read;
 
 pub use compact::CompactOptions;
 pub use fuse_sink::BloomIndexState;
